requirements.txt
amqp==5.1.1
aniso8601==7.0.0
asgiref==3.5.0
billiard==3.6.4.0
black==22.3.0
celery==5.2.6
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.3
django-audiofield==0.10.0
django-celery==3.1.17
django-cors-headers==3.11.0
django-countries==7.3.2
django-filter==21.1
django-graphql-jwt==0.3.4
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
kombu==5.2.4
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
promise==2.3
prompt-toolkit==3.0.29
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2022.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.2.0
vine==5.0.0
wcwidth==0.2.5
